+++ src/app/models/character.model.ts
