+++ src/app/tables/items/items.page.ts
@@ -2,7 +2,7 @@
 import { Component, OnDestroy, OnInit } from '@angular/core';
 import { Subscription } from 'rxjs';
 import { WaghamLoadingController } from '../../shared/wagham-loading-controller';
-import { Item, ItemTableRow } from '../../shared/models/items.model';
+import { Item, ItemTableRow } from '../../models/items.model';
 import { ItemService } from 'src/app/services/item.service';
 import { LinkCommand } from 'src/app/shared/commands/url-command/link-command';
