+++ src/app/tabs/character/character.page.ts
@@ -6,9 +6,9 @@ import { take, tap } from 'rxjs/operators';
 import { AuthService } from 'src/app/services/auth.service';
 import { IllegalStateError } from 'src/app/shared/illegal-state-error';
 import { WaghamLoadingController } from 'src/app/shared/wagham-loading-controller';
-import { Character } from '../../shared/models/character.model';
+import { Character } from '../../models/character.model';
 import { CharacterService } from '../../services/character.service';
-import { MSTable } from 'src/app/shared/models/msTable.model';
+import { MSTable } from 'src/app/models/msTable.model';
 
 @Component({
   selector: 'app-character',
